// WFX_Code/WFXWeb/WFX.Web/ClientApp/src/models/QCSearchModel.ts
export class QCSearchModel {
   productlist: {
    id: number;
    text: string;
  }

  stylelist: {
    id: number;
    text: string;
  }

  productfitlist: {
    id: number;
    text: string;
  }

  seasonlist: {
    id: number;
    text: string;
  }

  customerlist: {
    id: number;
    text: string;
  }

  solist: {
    id: number;
    text: string;
  }

  modulelist: {
    id: number;
    text: string;
  }

  linelist: {
    id: number;
    text: string;
  }

  polist: {
    id: number;
    text: string;
  }

  processlist: {
    id: number;
    text: string;
  }
}